VBoxSVGA3D: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@62948 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Graphics/shaderlib/shader_sm1.c
@@ -423,7 +423,7 @@ static const struct wined3d_sm1_opcode_info *shader_get_opcode(const struct wine
 
     while (opcode_table[i].handler_idx != WINED3DSIH_TABLE_SIZE)
     {
-        if ((code & WINED3DSI_OPCODE_MASK) == opcode_table[i].opcode
+        if ((code & WINED3DSI_OPCODE_MASK) == (DWORD)opcode_table[i].opcode
                 && shader_version >= opcode_table[i].min_version
                 && (!opcode_table[i].max_version || shader_version <= opcode_table[i].max_version))
         {
